Refactor ConsumeDataEventStreamsView component
